Add QuickGuidePanels-specs Enact-DCO-1.0-Signed-off-by: Jiye Kim (jiye.kim@lge.com)
